+++ perf-event/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Expose the `IOC_SET_BPF` ioctl as `Counter::set_bpf`.
+- Add `KProbe` and `UProbe` events.
 
 ## [0.6.0] - 2023-05-17
 ### Added